add reactif,echantilln

+++ src/main/java/com/example/techlab/entities/Reactif.java
@@ -4,7 +4,6 @@ import lombok.*;
 
 import javax.persistence.*;
 import java.time.LocalDate;
-import java.util.Date;
 import java.util.List;
 
 @Entity
@@ -26,7 +25,7 @@ public class Reactif {
     private String description;
 
     @Column(name = "quantite_reactif")
-    private double quantiteReactif;
+    private int quantiteReactif;
 
     @Column(name = "date_expiration")
     private LocalDate dateExpiration;
